% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a1rk.r
\name{a1rk}
\alias{a1rk}
\title{Rank of the A-1 homotopy group of a graph}
\usage{
a1rk(graph, componentwise = FALSE)
}
\arguments{
\item{graph}{The input graph. It will be treated as simple and undirected. A 
submodule handles the case that \code{graph} is connected, and a wrapper 
feeds this submodule the connected components of \code{graph}.}

\item{componentwise}{Logical, defaults to false. If true, returns a vector of
the A-1 ranks of the connected components of \code{graph}. If false,
returns the A-1 rank of \code{graph}, which is the sum of these ranks.}
}
\description{
Calculate the rank (the number of homotopy-inequivalent cycles) of the first 
discrete homotopy group of a graph.
}
\examples{
# cycle graph
g <- graph(c(1,2, 2,3, 3,4, 4,5, 5,1), directed = FALSE)
a1rk(g)

# two-cycle graph
h <- add_vertices(g, 4)
h <- add_edges(h, c(2,6, 6,7, 7,8, 8,9, 9,1))
a1rk(h)

# Platonic solids
a1rk(graph.famous("tetrahedron"))
a1rk(graph.famous("cubical"))
a1rk(graph.famous("dodecahedron"))
}

