% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-triad-census.r
\name{simple_triad_census}
\alias{simple_triad_census}
\title{Simple triad census}
\usage{
simple_triad_census(graph)
}
\arguments{
\item{graph}{The input graph. It can be directed or undirected but the
function will return the census for its undirected version.}
}
\description{
Calculate the triad census for an undirected graph by taking the appropriate 
sums from the triad census for a directed graph with the same edges (up to 
direction).
}

