% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-sc.r
\name{top_sc}
\alias{top_sc}
\title{Remove non-maximal simplices from a simplicial complex}
\usage{
top_sc(sc)
}
\arguments{
\item{sc}{The input simplicial complex, represented as a list of simplicies, 
each a vector of positive integers representing the nodes.}
}
\description{
Remove non-maximal simplices from a simplicial complex
}

