% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-to-graph.r
\name{sc_to_graph}
\alias{sc_to_graph}
\title{q-connectivity graph for a simplicial complex}
\usage{
sc_to_graph(sc, q)
}
\arguments{
\item{sc}{The input simplicial complex, represented as a list of simplicies, 
each a vector of positive integers representing the nodes.}

\item{q}{A positive integer, the minimal dimension of a face two simplices
must share in order to be connected.}
}
\description{
Given a simplicial complex and a dimension q, construct the q-connectivity 
graph, which is the graph whose nodes are the maximal simplices of the 
complex and whose edges link those pairs of simplices that share a face of 
dimension at least q.
}

